rename : 파일 이동 및 gitignore에 http 관련 파일 추가
